Started to rework deletion checks so no unintended database issues can arise.
We should really start working with referential integrity, otherwise this will become unmanageable soon.

That would also require things like Taxes be referred to by their IDs, not their names.

https://github.com/FOSSBilling/FOSSBilling/blob/ad94a9f3255801817f393cda36bc0a38796e1eb9/src/install/sql/structure.sql#L575

https://github.com/FOSSBilling/FOSSBilling/blob/ad94a9f3255801817f393cda36bc0a38796e1eb9/src/install/sql/structure.sql#L1365

+++ src/modules/Servicehosting/Api/Admin.php
@@ -227,6 +227,14 @@ class Admin extends \Api_Abstract
 
         $model = $this->di['db']->getExistingModelById('ServiceHostingServer', $data['id'], 'Server not found');
 
+        // check if server is not used by any service_hostings
+        $hostings = $this->di['db']->find('ServiceHosting', 'service_hosting_server_id = :cart_id',[':cart_id' => $data['id']]);
+        $count = count($hostings);
+        if ($count > 0)
+        {
+            throw new \Box_Exception('Hosting server is used by :count: service hostings', [':count:'=> $count], 704);
+        }
+
         return (bool) $this->getService()->deleteServer($model);
     }
 
@@ -330,6 +338,13 @@ class Admin extends \Api_Abstract
 
         $model = $this->di['db']->getExistingModelById('ServiceHostingHp', $data['id'], 'Hosting plan not found');
 
+        // check if hosting plan is not used by any service_hostings
+        $hostings = $this->di['db']->find('ServiceHosting', 'service_hosting_hp_id = :cart_id',[':cart_id' => $data['id']]);
+        $count = count($hostings);
+        if ($count > 0)
+        {
+            throw new \Box_Exception('Hosting plan is used by :count: service hostings', [':count:'=> $count], 704);
+        }
         return (bool) $this->getService()->deleteHp($model);
     }
